make this work with the new workday spreadsheet format
I tried to convert xlsx to iCal for my schedule this year, but the code was not working. I discovered that there are a two new columns and some rearrangements to the formatting of the spreadsheet. See my initial, temporary workaround: https://imgur.com/a/6KAf4fA

This fix changes which column the code looks at if it finds the new columns. This means that .xlsx in the old format still can get converted